% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binpL.R
\name{binpL}
\alias{binpL}
\title{Likelihood-based Confidence Interval for the Binomial}
\usage{
binpL(a, n, alpha, toler)
}
\arguments{
\item{a}{the number of successes.}

\item{n}{total number of trials.}

\item{alpha}{the significance level used, 1 - alpha interval calculated, default = 0.05.}

\item{toler}{the desired accuracy using optimise, default = 0.0001.}
}
\value{
$prob.val - MLE for probability from the data.

$succ.fail - number of successes and failures.

$conf.int - likelihood-based confidence interval.

$err.acc - error accuracy for optimize function.
}
\description{
This function allows you to calculate the usual confidence interval, but is likelihood-based.
It uses the optimize function to locate desired limits and their error.
}
\examples{
binpL(86,94,0.05)

}
\references{
Aitkin, M. et al (1989) Statistical Modelling in GLIM, Clarendon Press, ISBN : 978-0198522041

Cahusac, P.M.B. (2020) Evidence-Based Statistics, Wiley, ISBN : 978-1119549802

Royall, R. M. (1997). Statistical evidence: A likelihood paradigm. London: Chapman & Hall, ISBN : 978-0412044113

Edwards, A.W.F. (1992) Likelihood, Johns Hopkins Press, ISBN : 978-0801844430
}
\keyword{Likelihood-based;}
\keyword{binomial,}
\keyword{confidence}
\keyword{interval}
