% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_1way_ANOVA.R
\name{L_1way_ANOVA}
\alias{L_1way_ANOVA}
\title{Likelihood Supports for One-way independent samples ANOVA}
\usage{
L_1way_ANOVA(data, group, cont1=NULL, cont2=NULL)
}
\arguments{
\item{data}{a (non-empty) numeric vector of data values.}

\item{group}{an integer vector the same length as data, coding for k groups.}

\item{cont1}{first contrast, default = NULL.}

\item{cont2}{second contrast, default = NULL.}
}
\value{
$S.12c - corrected support for groups hypothesis versus null.

$S.12 - uncorrected support.

$df - degrees of freedom for ANOVA.

$S.cont.12 - support for contrast 1 versus contrast 2.

$gp.means - group means.
}
\description{
This function calculates supports for independent samples ANOVA. One support is
for the model of group means against the null (no grouping), and the other for 2 contrasts.
If the contrasts are not specified then it calculates a linear versus a
quadratic contrast. The corrected support is given for groups versus null, using
Akaike's correction (Hurvich & Tsai (1989)). No correction is necessary for the
two contrasts support since they both involve 1 parameter. Unequal group sizes can be used.
}
\examples{
# fitness example, p 81
dat <- c(7,	5,	9,	8,	3,	12,	10,	8,	7,	9,
5,	7,	7,	6,	4,	8,	12,	9,	7,	8,
3,	2,	7,	6,	8,	6,	5,	3,	4,	3,
4,	3,	3,	1,	2,	5,	7,	6,	8,	7)
gp <- as.factor(rep(1:4,each=10))
gp = gl(4,10,40, labels=c("6 hr",	"3 hr",	"Sports club",	"Video games"))
contrast1 <- c(-3, -1, 1, 3)  # linear
contrast2 <- c(1, -1, -1, 1)  # quadratic
contrast3 <- c(1, 1, -1, -1)
L_1way_ANOVA(dat,gp,contrast3, contrast1)

}
\references{
Cahusac, P.M.B. (2020) Evidence-Based Statistics, Wiley, ISBN : 978-1119549802

Hurvich CM, Tsai C-L. Regression and time series model selection in small
samples. Biometrika. 1989; 76(2):297–307.

Dixon P. The effective number of parameters in post hoc models. Behavior
Research Methods. 2013; 45(3):604–12.

Dixon P. The p-value fallacy and how to avoid it. Canadian Journal of
Experimental Psychology/Revue canadienne de psychologie expérimentale. 2003;
57(3):189–202.

Glover S, Dixon P. Likelihood ratios: a simple and flexible statistic for empirical
psychologists. Psychonomic Bulletin and Review. 2004; 11(5):791–806.
}
\keyword{ANOVA;}
\keyword{Likelihood;}
\keyword{contrasts}
\keyword{independent}
\keyword{samples}
\keyword{support;}
