% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_1way_RM_ANOVA.R
\name{L_1way_RM_ANOVA}
\alias{L_1way_RM_ANOVA}
\title{Likelihood Supports for One-way Repeated Measures ANOVA}
\usage{
L_1way_RM_ANOVA(dat, group, ID)
}
\arguments{
\item{dat}{a (non-empty) numeric vector of data values.}

\item{group}{a vector the same length as data, coding for k groups.}

\item{ID}{is an identifier for each case.}
}
\value{
$S.12 - support for groups means hypothesis versus null.

$df - degrees of freedom for ANOVA.

$gp.means - group means.
}
\description{
This function calculates support for repeated measures ANOVA. One support is
for the model of group means against the null (no grouping), and the other for 2 contrasts.
If the contrasts are not specified then it calculates a linear versus a
quadratic contrast. The corrected support is given for groups versus null, using
Akaike's correction (Hurvich & Tsai (1989)). No correction is necessary for the
two contrasts support since they both involve 1 parameter. Unequal group sizes can be used.
}
\examples{
# sleep data example, p 97
dat <- c(0.7,	-1.6,	-0.2,	-1.2,	-0.1,	3.4,	3.7,	0.8,	0,	2,
1.9,	0.8,	1.1,	0.1,	-0.1,	4.4,	5.5,	1.6,	4.6,	3.4,
1.5,	1.4,	0.0,	-0.7,	0.5,	5.1,	5.7,	1.5,	4.7,	3.5)
treat <- gl(3,10,30)
patients <- gl(10,1,30)

m=L_1way_RM_ANOVA(dat, treat, patients)
m

}
\references{
Cahusac, P.M.B. (2020) Evidence-Based Statistics, Wiley, ISBN : 978-1119549802

Dixon P. The effective number of parameters in post hoc models. Behavior
Research Methods. 2013; 45(3):604–12.

Dixon P. The p-value fallacy and how to avoid it. Canadian Journal of
Experimental Psychology/Revue canadienne de psychologie expérimentale. 2003;
57(3):189–202.

Glover S, Dixon P. Likelihood ratios: a simple and flexible statistic for empirical
psychologists. Psychonomic Bulletin and Review. 2004; 11(5):791–806.
}
\keyword{ANOVA}
\keyword{Likelihood;}
\keyword{measures}
\keyword{repeated}
\keyword{support;}
